Refactor: flatten scene structure, move and clean up all game/scene/map/particle/generator/blob components, fix all imports, remove obsolete AnimationLayer, and ensure clean build
