+++ src/App.tsx
@@ -1,15 +1,9 @@
 import "./globals.css";
-import { AnimationLayer } from "./components/animations/AnimationLayer";
-import { FlyingParticles } from "./components/animations/FlyingParticles";
-import { BlobContainer } from "./components/blob/BlobContainer";
-
+import { GameScene } from "./components/GameScene";
 import { GameHUD } from "./components/hud/GameHUD";
-import { MapGenerators } from "./components/map/MapGenerators";
-import { GameWorld } from "./components/GameWorld";
 import { useGame } from "./hooks/useGame";
 import { useCameraZoom } from "./hooks/useCameraZoom";
 import { useBlobSize } from "./hooks/useBlobSize";
-import Map from "./components/map/Map";
 import { getCurrentLevel } from "./game/systems/actions";
 
 function App() {
@@ -28,12 +22,6 @@ function App() {
 
   return (
     <div className="w-screen h-screen relative overflow-hidden">
-      {/* Game World Container - Everything that zooms */}
-      <GameWorld zoom={currentZoom}>
-        {/* Background Layer - z-index: 0 */}
-        <Map className="absolute inset-0 w-full h-full z-0" />
-      </GameWorld>
-
       {/* HUD Layer - Outside zoom container */}
       <GameHUD
         biomass={gameState.biomass}
@@ -45,31 +33,13 @@ function App() {
         blobSize={blobSize}
       />
 
-      {/* Animation Layer with BlobContainer and MapGenerators - Outside zoom container */}
-      <AnimationLayer>
-        {(addFloatingNumber) => (
-          <>
-            {/* Blob Layer - z-index: 70 (above generators) */}
-            <BlobContainer
-              id="main-blob"
-              biomass={gameState.biomass}
-              size={blobSize}
-              onBlobClick={handleBlobClick}
-              clickPower={gameState.clickPower}
-              addFloatingNumber={addFloatingNumber}
-            />
-            
-            <MapGenerators 
-              gameState={gameState}
-              blobSize={blobSize}
-              addFloatingNumber={addFloatingNumber}
-            />
-          </>
-        )}
-      </AnimationLayer>
-
-      {/* Particle System Layer - Outside zoom container, same as blob */}
-      <FlyingParticles gameState={gameState} currentLevel={currentLevel} />
+      {/* Game Scene Layer - Contains everything that zooms and the game world */}
+      <GameScene
+        gameState={gameState}
+        blobSize={blobSize}
+        onBlobClick={handleBlobClick}
+        zoom={currentZoom}
+      />
     </div>
   );
 }
